Update to add-to-cart-variations.js
I have updated the JS file with the latest Woocommerce Variations JS and amended it to work with the radio buttons. My issue before was that the Add to Cart button would not appear. After some delving into the code I realised that the JS was different. The above works for me with the radio buttons in Woocommerce VS: 2.4.6